requirements.txt
flask
Flask-SQLAlchemy
flask-login
python-dotenv
gunicorn